requirements.txt
crewai==1.12.2
langchain-google-genai==4.2.1
google-generativeai==0.8.6
pdfplumber==0.11.9
sentence-transformers==5.3.0
scikit-learn==1.8.0
python-dotenv
fastapi
uvicorn
python-multipart